refactor(dashboard): decompose keeper-detail.ts into flat keeper-detail-* modules
Split the 1315-line keeper-detail.ts into focused modules following the
fsm-hub-* flat naming precedent:

- keeper-detail-page.ts: KeeperDetailPage shell (state, effects, header)
- keeper-detail-body.ts: KeeperDetailBody grid (sections, panels, dialogs)
- keeper-detail-state.ts: selectedKeeper signal & navigation helpers
- keeper-detail-helpers.ts: refreshAfterRuntimeAction, diagnostic attention
- keeper-detail-primitives.ts: StrongSecondary, RuntimeBadge
- keeper-detail-hooks.ts: useKeeperComposite hook
- keeper-detail-alert-strip.ts: KeeperRuntimeAlertStrip
- keeper-detail-lifecycle.ts: KeeperLifecycleButtons, KeeperClearContextDialog
- keeper-detail-comms.ts: KeeperCommsPanel, PlaygroundReposPanel

keeper-detail.ts becomes a backward-compatible barrel re-export.
Fix missing keeperDisplayModel/keeperActivityDisplay imports in page shell.

tsc --noEmit clean for new files. vitest: 7 files, 99 tests passed.
